fix: move dto to data package
