+++ ee/backend/app/src/main/kotlin/io/tolgee/ee/data/DomainRequest.kt
@@ -1,0 +1,6 @@
+package io.tolgee.ee.data
+
+data class DomainRequest(
+  val domain: String,
+  val state: String,
+)
