+++ ee/backend/app/src/main/kotlin/io/tolgee/ee/data/SsoUrlResponse.kt
@@ -1,0 +1,5 @@
+package io.tolgee.ee.data
+
+data class SsoUrlResponse(
+  val redirectUrl: String,
+)
